cart_service/requirements.txt
fastapi
uvicorn
selenium
ddddocr
Pillow
lark-oapi
python-dotenv